pyproject.toml
[project]
name = "tinygrad_unet"
version = "0.0.1"
license = "MIT"
dependencies = [
    "Pillow",
    "tinygrad"
]

[tool.setuptools]
packages = [
    "tinygrad_unet.dataset",
    "tinygrad_unet.net",
    "tinygrad_unet.inference",
    "tinygrad_unet.util"
]

[tool.pyright]
extraPaths = ["tinygrad"]
# typeCheckingMode = "basic"  # Less strict for dynamic tinygrad
